Create MergeNodesInBetweenZeros
We were given a linked list with like this :
0 -> 1 -> 3 -> 0 -> 1 -> 2 -> 0 ->1 ->0
We were to merge the nodes between zeros, i.e, add the values of the nodes and insert a new node containing that values. We were also required to remove all the zeroes and return the head of the modified list.