Create DeletingNodeButWeDontHaveAccessToHead
We were to delete a node but we were not given the head to the linked list that contained that node. The node was not the head or the tail of the linked list, that is it lied somehwhere in the middle of the linked list. We were to ensure that the node's memory allocation is preserved.